python/requirements.txt
google-genai
python-dotenv
openai
ollama
pydantic
typing
typing-extensions
dataclasses
langgraph
langchain-ollama
